+++ src/common/actions/File.js
@@ -34,7 +34,7 @@ var action = {
   name: _('Open File'),
 
   // This is also used in the add-new-item-popover.
-  icon: 'text-x-generic',
+  icon: 'flypie-action-file-symbolic-#8a3',
 
   // Translators: Please keep this short.
   // This is the (short) description shown in the add-new-item-popover.