+++ src/common/menus/FrequentlyUsed.js
@@ -44,7 +44,7 @@ var menu = {
   name: _('Frequently Used'),
 
   // This is also used in the add-new-item-popover.
-  icon: 'emblem-favorite',
+  icon: 'flypie-menu-frequently-used-symbolic-#b59',
 
   // Translators: Please keep this short.
   // This is the (short) description shown in the add-new-item-popover.